requirements.txt
Django>=4.2.0,<4.3.0
psycopg2-binary>=2.9.0
django-tailwind>=3.0.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
Pillow>=9.0.0
pytest
django
pytest-django
gunicorn>=20.1.0
whitenoise>=6.0.0
python-decouple>=3.6 